+++ lib/src/rspc/build.dart
@@ -38,7 +38,7 @@ void compileFile(String sourceName, {String destinationName, bool verbose : fals
     print("Compile ${source.path} to ${i > 0 ? dest.path.substring(i) : dest.path}");
   }
   
-  source.readAsString(encoding).then((text) {
+  source.readAsString(encoding: encoding).then((text) {
     final out = dest.openWrite(encoding: encoding);
     try {
       compile(text, out, sourceName: sourceName,